refactor: app code package structure for modularization, move auth files into auth package
