+++ code/app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".ProfileActivity" />
+        <activity android:name=".userProfile.ProfileActivity" />
     </application>
 
 </manifest>
