+++ code/app/src/main/java/com/hamidat/nullpointersapp/MainActivity.java
@@ -11,6 +11,8 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
+import com.hamidat.nullpointersapp.userProfile.ProfileActivity;
+
 public class MainActivity extends AppCompatActivity {
 
     @Override
