--- code/app/src/main/java/com/hamidat/nullpointersapp/SignUpActivity.java
@@ -1,6 +1,0 @@
-package com.hamidat.nullpointersapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class SignUpActivity extends AppCompatActivity {
-}
