+++ code/app/src/main/res/layout/activity_signup.xml
@@ -24,7 +24,7 @@
         app:layout_constraintTop_toBottomOf="@+id/etSignupUsername" />
 
     <Button
-        android:id="@+id/btnNext"
+        android:id="@+id/signUpButton"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:backgroundTint="@color/primaryColor"
